+++ apps/backend/src/libs/modules/database/libs/enums/database-table-name.enum.ts
@@ -3,6 +3,8 @@ const DatabaseTableName = {
 	MIGRATIONS: "migrations",
 	PLAN_DAYS: "plan_days",
 	PLANS: "plans",
+	QUESTION_OPTIONS: "question_options",
+	QUESTIONS: "questions",
 	TASKS: "tasks",
 	USERS: "users",
 } as const;