+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -5,6 +5,8 @@ import { authController } from "~/modules/auth/auth.js";
 import { feedbackController } from "~/modules/feedbacks/feedbacks.js";
 import { planDayController } from "~/modules/plan-days/plan-days.js";
 import { planController } from "~/modules/plans/plans.js";
+import { quizQuestionContolller } from "~/modules/quiz-questions/quiz-question.js";
+import { quizContolller } from "~/modules/quiz/quiz.js";
 import { taskController } from "~/modules/tasks/tasks.js";
 import { userController } from "~/modules/users/users.js";
 
@@ -20,6 +22,8 @@ const apiV1 = new BaseServerApplicationApi(
 	...planDayController.routes,
 	...taskController.routes,
 	...feedbackController.routes,
+	...quizQuestionContolller.routes,
+	...quizContolller.routes,
 );
 const serverApplication = new BaseServerApplication({
 	apis: [apiV1],