+++ apps/frontend/src/assets/img/shared/illustrations/layouts/layouts.img.tsx
@@ -1,4 +1,5 @@
-export { default as ExampleColourful } from "./example-colourful.svg";
-export { default as ExampleMinimal } from "./example-minimal.svg";
-export { default as ExampleMotivating } from "./example-motivating.svg";
-export { default as ExampleWithRemarks } from "./example-with-remarks.svg";
+export { default as ExampleColourful } from "./example-colourful.png";
+export { default as ExampleMinimal } from "./example-minimal.png";
+export { default as ExampleMotivating } from "./example-motivating.png";
+export { default as ExampleWithRemarks } from "./example-with-remarks.png";
+export { default as ExampleWithRemarksBig } from "./example-with-remarks_plan_big.png";