+++ apps/frontend/src/libs/enums/app-route.enum.ts
@@ -1,7 +1,11 @@
 const AppRoute = {
+	CHOOSE_STYLE: "/choose-style",
 	DASHBOARD: "/dashboard",
 	NOT_FOUND: "*",
 	PLAN: "/plan",
+	PROFILE: "/profile",
+	QUIZ: "/quiz",
+	QUIZ_QUESTIONS: "/quiz/questions",
 	ROOT: "/",
 	SIGN_IN: "/sign-in",
 	SIGN_UP: "/sign-up",