+++ apps/frontend/src/libs/enums/categories-colors.enum.ts
@@ -1,0 +1,10 @@
+const CategoriesColors = {
+	CARD_CYAN: "card-cyan",
+	CARD_GRAY: "card-gray",
+	CARD_PEACH: "card-peach",
+	CARD_PINK: "card-pink",
+	CARD_PURPLE: "card-purple",
+	CARD_YELLOW: "card-yellow",
+} as const;
+
+export { CategoriesColors };