+++ apps/frontend/src/routes.tsx
@@ -3,7 +3,14 @@ import { type RouteObject } from "react-router-dom";
 import { Dashboard, DashboardWrapper } from "./libs/components/components.js";
 import { AppRoute, RouteAccess } from "./libs/enums/enums.js";
 import { Plan } from "./pages/dashboard-wrapper-mock/components/components.js";
-import { Auth, Home, NotFound } from "./pages/pages.js";
+import {
+	Auth,
+	ChooseStyle,
+	Home,
+	NotFound,
+	QuestionFlow,
+	Quiz,
+} from "./pages/pages.js";
 
 type CustomRouteObject = RouteObject & { handle: RouteHandle };
 
@@ -28,6 +35,16 @@ const routes: CustomRouteObject[] = [
 		path: AppRoute.SIGN_UP,
 	},
 	{
+		element: <QuestionFlow />,
+		handle: { access: RouteAccess.PUBLIC },
+		path: AppRoute.QUIZ_QUESTIONS,
+	},
+	{
+		element: <Quiz />,
+		handle: { access: RouteAccess.PUBLIC },
+		path: AppRoute.QUIZ,
+	},
+	{
 		children: [
 			{
 				element: <Dashboard />,
@@ -43,6 +60,11 @@ const routes: CustomRouteObject[] = [
 		element: <DashboardWrapper />,
 		handle: { access: RouteAccess.AUTHENTICATED },
 		path: AppRoute.ROOT,
+	},
+	{
+		element: <ChooseStyle />,
+		handle: { access: RouteAccess.AUTHENTICATED },
+		path: AppRoute.CHOOSE_STYLE,
 	},
 	{
 		element: <NotFound />,