+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -3,6 +3,8 @@ const APIPath = {
 	FEEDBACKS: "/feedbacks",
 	PLAN_DAYS: "/plan-days",
 	PLANS: "/plans",
+	QUIZ: "/quiz",
+	QUIZ_QUESTIONS: "/quiz-questions",
 	TASKS: "/tasks",
 	USERS: "/users",
 } as const;