+++ packages/shared/src/modules/quiz/quiz.ts
@@ -1,0 +1,19 @@
+export {
+	QuizApiPath,
+	QuizIndexes,
+	QuizQuestionFormat,
+} from "./libs/enums/enums.js";
+export {
+	type QuestionDto,
+	type QuestionOptionDto,
+	type QuestionType,
+	type QuizAnswer,
+	type QuizAnswersRequestDto,
+	type QuizAnswersResponseDto,
+	type QuizCategoryType,
+	type QuizQuestionsResponseDto,
+} from "./libs/types/types.js";
+export {
+	QuizAnswersSchema as QuizAnswersValidationSchema,
+	QuizStateSchema as QuizStateValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";