+++ packages/shared/src/modules/users/users.ts
@@ -11,6 +11,7 @@ export {
 	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
+	type UserUpdateRequestDto,
 } from "./libs/types/types.js";
 export {
 	type SignUpFormValidationSchema,